Move artifacts resources to own folder/package
