+++ pkg/artifactory/datasource/artifact/datasource_artifactory_fileinfo.go
@@ -1,4 +1,4 @@
-package datasource
+package artifact
 
 import (
 	"context"
